Fixes #12 Add time to watch
Added optional interval parameter for further improvements in the later
commit. Calculated the terminal with and outputted information about
the command the update interval along with the current time. Changed to
use usleep for microseconds. The watch trait no longer looks for a
‘watch’ input option and continues forever. Also added new watch
command as a shortcut with added option to specify the refresh interval.